% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D2.R
\name{calculateD2}
\alias{calculateD2}
\title{Calcuate Mahalanobis D2 Measures}
\usage{
calculateD2(binary_trait_data, deltamin = 0.01)
}
\arguments{
\item{binary_trait_data}{a data frame with binary data}

\item{deltamin}{will replace any negative distance}
}
\description{
Based on Lyle W. Konigsberg scripts (tdistR.zip) http://lylek.ucoz.org/index.html
with some additional cleaning and regularisation
}
