% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmd.R
\name{calculateMMD}
\alias{calculateMMD}
\title{Calcuate MMD Measures}
\usage{
calculateMMD(M_n, M_p, thetadiff, theta)
}
\arguments{
\item{M_n}{Matrix with number of observations for each trait and each site}

\item{M_p}{Matrix with proportions for each trait and each site}

\item{thetadiff}{Function for correction of the MMD}

\item{theta}{Function for MMD calculations}
}
\value{
a list with tree matrices: MMDMatrix - distance, SDMatrix - standard deviation, SigMatrix - significance
}
\description{
Base on Arkadiusz Soltysiak ,,Technical Note: An R script for Smith's Mean Measure of Divergence''
}
