% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_Mn_Mp}
\alias{get_Mn_Mp}
\title{Extracts Mn and Mp matricex from binary data}
\usage{
get_Mn_Mp(binary_trait_data)
}
\arguments{
\item{binary_trait_data}{matrix with binary data}
}
\value{
List with two matrices for Mn and Mp
}
\description{
Extracts Mn and Mp matricex from binary data
}

