% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClust.R
\name{getClust}
\alias{getClust}
\title{Create a clustering diagram for given distance matrix}
\usage{
getClust(mat)
}
\arguments{
\item{mat}{distance matrix}
}
\value{
ggplot2 plot
}
\description{
Create a clustering diagram for given distance matrix
}
