% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCutoffStats.R
\name{getCutoffStats}
\alias{getCutoffStats}
\title{Create a cutoff statistics}
\usage{
getCutoffStats(df)
}
\arguments{
\item{df}{dataframe with variables}
}
\value{
data frame with summary statistics
}
\description{
Create a cutoff statistics
}
