% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCzekanowski.R
\name{getCzekanowski}
\alias{getCzekanowski}
\title{Create a Czekanowski diagram (MaCzek) for given distance matrix}
\usage{
getCzekanowski(mat)
}
\arguments{
\item{mat}{distance matrix}
}
\value{
ggplot2 plot
}
\description{
Based on Czekanowski diagrams as described in http://antropologia.uw.edu.pl/MaCzek/maczek.html
}
