% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDS.R
\name{getMDS}
\alias{getMDS}
\title{Create a MDS diagram for given distance matrix}
\usage{
getMDS(mat)
}
\arguments{
\item{mat}{distance matrix}
}
\value{
ggplot2 plot
}
\description{
Create a MDS diagram for given distance matrix
}

