% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummaryStatistics.R
\name{getSummaryStatistics}
\alias{getSummaryStatistics}
\title{Create a summary statistics}
\usage{
getSummaryStatistics(df)
}
\arguments{
\item{df}{dataframe with variables}
}
\value{
data frame with summary statistics
}
\description{
Create a summary statistics
}
