% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Read Data From Excel File}
\usage{
loadData(file)
}
\arguments{
\item{file}{xlsx file with data}
}
\value{
a data frame with data
}
\description{
First three collumns in the file should contain ID, SITE and SEX,
following columns shall contain trait scores in ordered scale.
Moreover, it is assumed, that the second row is names THRESHOLD
and contains cut off values.
}

