% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theta_Anscombe}
\alias{theta_Anscombe}
\title{Anscombe transformation}
\usage{
theta_Anscombe(n, p)
}
\arguments{
\item{n}{number of observations}

\item{p}{proportions}
}
\value{
theta MMD
}
\description{
Anscombe transformation
}
