% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theta_Freeman}
\alias{theta_Freeman}
\title{Freeman & Tukey transformation}
\usage{
theta_Freeman(n, p)
}
\arguments{
\item{n}{number of observations}

\item{p}{proportions}
}
\value{
theta MMD
}
\description{
Freeman & Tukey transformation
}

