% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{thetadiff_Grewal}
\alias{thetadiff_Grewal}
\title{Grewal correction}
\usage{
thetadiff_Grewal(nA, pA, nB, pB, theta)
}
\arguments{
\item{nA}{number of observations in the first group}

\item{pA}{proportions in the first group}

\item{nB}{number of observations in the second group}

\item{pB}{proportions in the second group}

\item{theta}{function for theta MMD calculations}
}
\value{
corrected theta MMD
}
\description{
Grewal correction
}
