% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D2.R
\name{calculateD2}
\alias{calculateD2}
\title{Calcuate Mahalanobis D2 Measures}
\usage{
calculateD2(binary_trait_data)
}
\description{
Based on Lyle W. Konigsberg scripts (tdistR.zip) http://lylek.ucoz.org/index.html
with some additional cleaning and regularisation
}
