% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDS.R
\name{getMDS}
\alias{getMDS}
\title{Create a MDS diagram for given distance matrix}
\usage{
getMDS(mat, metric = TRUE)
}
\arguments{
\item{mat}{distance matrix}

\item{metric}{metric (cmdscale) or non metric (isoMDS) scaling?}
}
\value{
ggplot2 plot
}
\description{
Create a MDS diagram for given distance matrix
}

